feat(modules): Add a new wizard implementation
This Wizard2 implementation is a controlled component with a
more straightforward API for the consumer. The parent consumer
can now navigate through the wizard without manging the state
within the parent. Instead, the indexing is handled by the
Wizard2 component itself.

Parent has to use a ref to access and invoke the handlers exposed
by the Wizard2. Check the `useImperativeHandle` to see exposed
handlers.